app/requirements.txt
fastapi==0.115.11
httpx==0.28.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-dotenv==1.0.1
python-multipart==0.0.20
uvicorn==0.34.0
sqlalchemy==2.0.39
greenlet==3.1.1
asyncpg==0.30.0
alembic==1.15.1
python-jose==3.4.0
cryptography==44.0.2
bcrypt==4.3.0
passlib==1.7.4
redis==4.6.0
celery==5.4.0
fastapi-cache2==0.2.2
